其他厂家接入FSN功能

+++ KyModel/Models/ky_atm.cs
@@ -5,7 +5,7 @@ namespace KyModel.Models
 {
     public partial class ky_atm
     {
-        public long kId { get; set; }
+        public int kId { get; set; }
         public string kATMNumber { get; set; }
         public string kATMAddress { get; set; }
         public int kNodeId { get; set; }
